server: Reorganise model and mutator types a bit
